requirements.txt
# 激光毁伤效能分析软件 - 核心依赖包
# 基于ANSYS 2021 R1，适用于conda环境jg

# 核心科学计算包
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
pandas>=1.3.0

# PyANSYS集成 - 核心包
ansys-mapdl-core>=0.65.0
ansys-fluent-core>=0.15.0
ansys-dpf-core>=0.8.0
pyansys>=0.61.0

# GUI框架
PyQt5>=5.15.0

# 可视化支持
pyvista>=0.37.0
vtk>=9.2.0

# 系统工具
pyyaml>=6.0
loguru>=0.6.0
click>=8.1.0
tqdm>=4.64.0

# 数据处理
openpyxl>=3.0.0
xlsxwriter>=3.0.0

# 测试工具
pytest>=7.0.0

# Windows特定
pywin32>=227; sys_platform == "win32"
